create subclasses to not repeat code
